Moved Fragments to "fragments" resource directory, and AppExecutors to "util" resource directory
