+++ app/src/main/java/com/example/android/bakingguru/RecipeDetailActivity.java
@@ -1,15 +1,13 @@
 package com.example.android.bakingguru;
 
-import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
-import androidx.appcompat.widget.Toolbar;
 import androidx.fragment.app.FragmentManager;
 
 import android.content.Intent;
 import android.os.Bundle;
-import android.os.PersistableBundle;
-import android.view.Menu;
 
+import com.example.android.bakingguru.fragments.RecipeDetailFragment;
+import com.example.android.bakingguru.fragments.RecipeListFragment;
 import com.example.android.bakingguru.util.AppUtil;
 
 public class RecipeDetailActivity extends AppCompatActivity {
