+++ app/src/main/res/layout/fragment_recipe_detail.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:layout_margin="20dp"
-    tools:context=".RecipeDetailFragment">
+    tools:context=".fragments.RecipeDetailFragment">
 
     <LinearLayout
         android:orientation="vertical"
