+++ app/src/main/res/layout/fragment_recipe_list.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".RecipeListFragment">
+    tools:context=".fragments.RecipeListFragment">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/rv_recipes"
